refactoring gui components
- update the old acknowledgement string
- improve error handling in some really old gui code
- refactor OUT of the Communique "god-class" two components for the text area and the timer progress bar
- finally get the timer progress bar to synchronise with the "stop" button
- delete the daemon mode class
- refactor communique document listener out of Communique.java and use a timer to make it wait a second after typing stops to trigger save commands

+++ src/com/git/ifly6/communique/ngui/CommuniqueSendDialog.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 ifly6
+ * Copyright (c) 2024 ifly6
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this class file and associated
  * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
@@ -162,7 +162,7 @@ public class CommuniqueSendDialog extends JDialog {
 	}
 
 	private String estimateTime(int count, int delay) {
-		int seconds = Math.round(count * (int) (delay / 1000));
+		int seconds = Math.round(count * ((float) delay / 1000));
 		return CommuniqueUtilities.time(seconds);
 	}
